fix: Find user by email
This change fixes the issue when the user is locally registered in the 
db and uses Google to login. The returned user would be the one from the 
database which can potentially have admin rights. If the user wasn't 
registered locally, then the temporary user gets deleted after the 
logout.